+++ pages/Common.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export function Link(props: LinkProps) {
+  return (
+    <a className="text-green hover:underline" href={props.link}>{props.children}</a>
+  );
+}
+
+interface LinkProps extends React.PropsWithChildren {
+  link: string;
+}
