+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from "next"
 import { useState } from "react"
-import Navbar from "./Left"
-import Content from "./Right"
+import Navbar from "./Navbar"
+import Content from "./Content"
 
 export enum CategoryState {
   ABOUT,
